Scatter functionality speaded.
This is an important commit.

Two new classes were created the ScatterSimulation and
SingleScatterSimulation. Essentially they will do the job the old
ScatterEstimationByBin did in a more modular, object oriented way.

While ScatterEstimationByBin will hold the general data handling, the
creation of maskes and the core iterative process.

Currently the code is incomplete.