Process filesignature.
There is a weird point, as in order to get to the version I need to skip to byte 32 and read it as big endian. The header kind and header size are read correctly, as little endians in the right place.  In the print_header I cannot any relevant process.